requirements.txt
absl-py==2.3.1
altair==5.5.0
astunparse==1.6.3
attrs==25.3.0
audioread==3.0.1
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
flatbuffers==25.2.10
fonttools==4.59.0
gast==0.6.0
gitdb==4.0.12
gitpython==3.1.45
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
idna==3.10
jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keras==3.10.0
kiwisolver==1.4.8
lazy-loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.44.0
markdown==3.8.2
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml-dtypes==0.5.1
msgpack==1.1.1
namex==0.1.0
narwhals==1.48.1
noisereduce==3.0.3
numba==0.61.2
numpy==2.1.3
opt-einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
protobuf==5.29.5
pyarrow==21.0.0
pyaudio==0.2.14
pycparser==2.22
pydeck==0.9.1
pydub==0.25.1
pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rich==14.1.0
rpds-py==0.26.0
scikit-learn==1.7.1
scipy==1.16.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
soundfile==0.13.1
soxr==0.5.0.post1
streamlit==1.47.1
streamlit-audiorecorder==0.0.6
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
